\name{MetSizeR-package}
\Rdversion{0.99}
\alias{MetSizeR-package}
\docType{package}
\title{
MetSizeR Graphical User Interface
}
\description{
Start the Graphical User Interface for MetSizeR. \cr
MetSizeR can be used to determine sample sizes for both targeted and NMR metabolomic experiments even when experimental pilot data is not available.
}
\details{
\tabular{ll}{
Package: \tab MetSizeR\cr
Type: \tab Package\cr
Version: \tab 0.99\cr
Date: \tab 2013-09-08\cr
License: \tab GPL (>= 2) \cr
LazyLoad: \tab yes\cr
}
The manual and example can be found with \code{vignette("MetSizeR")}.
}
\author{
Gift Nyamundanda \email{gift.nyamundanda@ucd.ie}, Isobel Claire Gormley, Yue Fan, William M Gallagher, Lorraine Brennan

Maintainer: Gift Nyamundanda \email{gift.nyamundanda@ucd.ie}
}

\keyword{ package }
