% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{partition}
\alias{partition}
\title{Generate n random splits of a data frame}
\usage{
partition(df, n, probs)
}
\arguments{
\item{df}{a data frame}

\item{n}{Number of observations to be grouped}

\item{probs}{A named vector of probabilities for groups}
}
\description{
Generate n random splits of a data frame
}

