% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{random_group}
\alias{random_group}
\title{Split data into random groups}
\usage{
random_group(n, probs)
}
\arguments{
\item{n}{Number of observations to be grouped}

\item{probs}{A named vector of probabilities for groups}
}
\description{
Split data into random groups
}

