% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sample_dist}
\alias{sample_dist}
\title{Sample from the null distribution}
\usage{
sample_dist(x, n_sims, n1, n2, conf_idx)
}
\arguments{
\item{x}{a numeric dataframe}

\item{n_sims}{number of simulations}

\item{n1}{number of samples in treatment A}

\item{n2}{number of samples in treatment B}
}
\description{
Computes pooled standard deviations and means
from simulated samples
}
\examples{
# simulate 100 split samples of the iris dataset and compute
sample_dist(iris[-5], 100, 10, 9, 3)
}

