% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pilot_data.R
\name{sim_DPPCA_data_with_pilot}
\alias{sim_DPPCA_data_with_pilot}
\title{Simulate data for DPPCA when pilot data is present}
\usage{
sim_DPPCA_data_with_pilot(n, n_pilot_vars, n_latent_dims)
}
\arguments{
\item{n}{number of samples to be simulated}

\item{n_pilot_vars}{number of variables in the pilot data}

\item{n_latent_dims}{number of latent dimension}
}
\description{
Simulate data for DPPCA when pilot data is present
}

