% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pilot_data.R
\name{sim_PPCA_data_with_pilot}
\alias{sim_PPCA_data_with_pilot}
\title{Simulate data for PPCA when pilot data is present}
\usage{
sim_PPCA_data_with_pilot(n_samps, n_pilot_vars, n_covars, n_latent_dims,
  ppcca_coefs)
}
\arguments{
\item{n_samps}{number of samples}

\item{n_pilot_vars}{the number of variables in the pilot data}

\item{n_covars}{the number of covariates}

\item{n_latent_dims}{the number of latent dimensions (set as 2 always)}
}
\description{
Simulate data for PPCA when pilot data is present
}

