% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pilot_data.R
\name{sim_PPCCA_data_with_pilot}
\alias{sim_PPCCA_data_with_pilot}
\title{Simulate data for PPCCA when pilot data is present}
\usage{
sim_PPCCA_data_with_pilot(n, n_pilot_vars, n_latent_dims, ppca_sigma,
  ppca_loadings, pilot_var_means)
}
\arguments{
\item{n}{number of samples}

\item{n_latent_dims}{number of latent dimensions}

\item{ppca_sigma}{the posterior mode estimate of the variance
of the error terms (MetabolAnalyze docs)}

\item{ppca_loadings}{a loadings matrix}

\item{pilot_var_means}{the means of the variables from the pilot data}
}
\description{
Simulate data for PPCCA when pilot data is present
}

