% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardize}
\alias{standardize}
\title{Scale atomic vector to lie between 0 and 1}
\usage{
standardize(x)
}
\arguments{
\item{x}{A numeric vector}
}
\description{
This function takes a vector of data and scales it to
lie between the values 0 and 1. It is intended as a helper function
and can be used with \code{map}
}
\examples{
# Standardize a vector
x <- rnorm(100, 3, 5)
standardize(x)

# Standardize the columns of a data frame
iris \%>\%
 select(-Species) \%>\%
 purrr::dmap(standardize)
}

