% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetSizeR.R
\name{MetSizeR}
\alias{MetSizeR}
\title{MetsizeR Launcher}
\usage{
MetSizeR()
}
\description{
Launches the Shiny application for the MetsizeR functionality. The
interactive application allows estimation of the sample size required for a
metabolomic experiment to achieve a desired statistical power.
}
\references{
G. Nyamundanda, I. C. Gormley, Y. Fan, W. M. Gallagher and L. Brennan, BMC Bioinformatics, 2013, 14, 338.
}
